requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
langchain
openai
langchain_openai
transformers
opencv-python
tensorflow
tf-keras
pillow
sentence-transformers
python-dotenv
cohere
langchain_cohere
nltk
pypdf2
python-multipart
google-generativeai
streamlit
faiss-cpu